Update: Bundle June 24, 2021

+++ app/models/domain/wallet-support/WalletSupport.scala
@@ -5,11 +5,11 @@ import java.time.Instant
 import play.api.libs.json._
 
 object Coin extends utils.CommonImplicits {
-	def apply(address: Option[String], currency: String, amount: Double): Coin = new Coin(address, currency, amount)
-	def apply(currency: String, amount: Double): Coin = Coin(None, currency, amount)
+	def apply(address: Option[String], currency: String, amount: BigDecimal): Coin = new Coin(address, currency, amount)
+	def apply(currency: String, amount: BigDecimal): Coin = Coin(None, currency, amount)
 	def apply(currency: String): Coin = Coin(currency, 0)
 }
-case class Coin(address: Option[String], currency: String, amount: Double) {
+case class Coin(address: Option[String], currency: String, amount: BigDecimal) {
 	def toJson(): JsValue = Json.toJson(this)
 }
 
@@ -21,7 +21,7 @@ case class CoinDeposit(txHash: String, issuer: Coin, receiver: Coin) {
 }
 
 object CoinWithdraw extends utils.CommonImplicits
-case class CoinWithdraw(receiver: Coin, gasPrice: Double) {
+case class CoinWithdraw(receiver: Coin, gasPrice: BigDecimal) {
 	def toJson(): JsValue = Json.toJson(this)
 	def toWalletHistory(txHash: String, id: UUID, txType: String, data: CryptoJsonRpcHistory): UserAccountWalletHistory =
 		new UserAccountWalletHistory(txHash, id, receiver.currency, txType, data, Instant.now)
